Create weighted-mean
Objective 
Practice calculating a weighted mean.

Task 
Given an array of integers and an array representing the respective weights of the given elements, calculate and print the weighted mean of of the elements. Your answer should be rounded to a scale of 1 decimal place.

Input Format
The first line contains an integer denoting the number of elements in arrays  and . 
The second line contains space-separated integers describing the respective elements of array . 
The third line contains space-separated integers describing the respective elements of array .

Output Format
Print the weighted mean on a new line. Your answer should be rounded to a scale of  decimal place (i.e., format).